requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests>=2.32.5,<3.0.0
unstructured[docx,pdf,pptx]==0.21.5
chromadb==0.4.24
ollama